msm: vidc: Handle the dangling sessions in hfi sessions list
When HW returns SYS_ERROR driver informs this error to all the
clients. Some clients might not close them immediately. But the
session_list is initialized when new video instance is started.
So the unclosed sessions are dangling in the hfi sessions list.
These need to be deleted properly from the list.

CRs-Fixed: 774249
Change-Id: I82f21659b0ebad383145cb13679faf30b9aa93e0
Signed-off-by: Praneeth Paladugu <ppaladug@codeaurora.org>